+++ src/frontend/src/flows/addDevice/addDeviceSuccess.json
