+++ src/frontend/src/flows/addDevice/manage/addDevice.ts
@@ -6,7 +6,7 @@ import { displayError } from "$src/components/displayError";
 import { withLoader } from "$src/components/loader";
 import { AuthenticatedConnection } from "$src/utils/iiConnection";
 import { isNullish } from "@dfinity/utils";
-import { addDeviceSuccess } from "./addDeviceSuccess";
+import { addDeviceSuccess } from "../addDeviceSuccess";
 import { addFIDODevice } from "./addFIDODevice";
 import { pollForTentativeDevice } from "./pollForTentativeDevice";
 import { verifyTentativeDevice } from "./verifyTentativeDevice";
