+++ src/showcase/src/pages/addDeviceSuccess.astro
@@ -7,7 +7,7 @@ import Screen from "../layouts/Screen.astro";
     import { toast } from "$src/components/toast";
     import { i18n } from "../i18n";
     import { chromeDevice } from "../constants";
-    import { addDeviceSuccessPage } from "$src/flows/addDevice/manage/addDeviceSuccess";
+    import { addDeviceSuccessPage } from "$src/flows/addDevice/addDeviceSuccess";
 
     addDeviceSuccessPage({
       i18n,
